UPGRADING instructions updated to work
I tried the draft upgrade instructions using the Conjur quick start
docker compose environment, but was unable to be successful updating the Conjur
container following them.

In this commit, I update the draft instructions to a workflow I was able to
demonstrate working with the Conjur quick start.

I am still unable to validate the custom upgrade instructions for v1.8.0.